% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafs.R
\name{grafica_barras}
\alias{grafica_barras}
\title{Gráfica de barras}
\usage{
grafica_barras(
  data,
  x,
  y,
  titulo = "Mi gráfico de barras bonito",
  subtitulo = "",
  titulo_x = "Eje X",
  titulo_y = "Eje Y",
  caption = "Fuente: Datos generados por mí",
  color_barras = "steelblue",
  tam_title = 16,
  tam_subtitulo = 14,
  tam_fuente = 12,
  tam_titulo_ejes = 10,
  tam_etiquetas_ejes = 10
)
}
\arguments{
\item{data}{Un data.frame}

\item{x}{Variable en el eje x}

\item{y}{Variable en el eje y}

\item{titulo}{Título del gráfico}
}
\value{
Un objeto ggplot
}
\description{
Gráfica de barras
}
